# backend/src/main/resources/openapi/API_REST.yaml
openapi: 3.0.3
info:
  title: Simulador MVP API
  version: 1.0.0
  description: |
    API REST del Simulador — plataforma de simulación empresarial competitiva paraguaya.

    ## Autenticación
    Todas las rutas excepto `/auth/login` y `/auth/refresh` requieren `Authorization: Bearer {access_token}`.
    El access token expira en **15 minutos**. Usar `POST /auth/refresh` con el refresh token (válido **30 días**) para obtener uno nuevo.

    ## Autorización
    Tres roles: `MODERADOR`, `JUGADOR`, `ADMIN_PLATAFORMA`.
    - `MODERADOR`: gestiona competencias que creó.
    - `JUGADOR`: participa en equipos, registra decisiones.
    - `ADMIN_PLATAFORMA`: gestiona catálogos (rubros, parámetros, eventos).

    ## Paginación
    Endpoints de listado aceptan `page` (default 0) y `size` (default 20, max 100). Responden con metadata.

    ## Errores
    Formato uniforme (RFC 7807 Problem Details) con `status`, `type`, `title`, `detail`, `errors[]` (campo a campo).

    ## Moneda
    Todos los montos se expresan en **guaraníes como enteros** (sin centavos).

    ## Tiempo real
    WebSocket en `/ws/competencias/{codigo}` para push de eventos (trimestre cerrado, evento disparado, resultados listos).

  contact:
    name: Equipo Simulador
    email: dev@simulador.py
servers:
  - url: https://api.simulador.py/v1
    description: Producción
  - url: https://api-stg.simulador.py/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local

...
paths:
...
  /equipos/{equipo_id}/trimestres/{trimestre_id}/bot/regenerar:
    post:
      tags: [Decisiones]
      operationId: regenerarDecisionBot
      summary: Re-ejecuta la heurística del bot y sobrescribe la decisión actual
      description: |
        Sólo aplica a equipos con `tipo='BOT'` en trimestres `ABIERTO_DECISIONES`.
        El moderador puede usar este endpoint si quiere "re-tirar" la decisión
        de un bot (por ejemplo después de calibrar coeficientes) sin tener que
        reabrir la decisión manualmente.

        Errores:
        - `403` si el caller no es el moderador de la competencia.
        - `404` si el equipo o el trimestre no existen.
        - `409` si el trimestre no está `ABIERTO_DECISIONES`.
        - `422` si el equipo no es de tipo `BOT`.
      parameters:
        - $ref: '#/components/parameters/EquipoIdNested'
        - $ref: '#/components/parameters/TrimestreIdNested'
      responses:
        '200':
          description: Decisión regenerada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Decision'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ------------------------------------------------------------------
  # EVENTOS EN COMPETENCIA
  # ------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ----- PARAMS -----
  parameters:
...
    EquipoIdNested:
      name: equipo_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    TrimestreIdNested:
      name: trimestre_id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  # ----- RESPONSES -----
  responses:
...
    Forbidden:
      description: Sin permisos para esta operación
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflicto de estado (ej. transición inválida, unicidad)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Error de validación de entrada
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemValidacion'

  # ----- SCHEMAS -----
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
          example: 'https://simulador.py/errors/validation'
        title:
          type: string
          example: 'Validación fallida'
        status:
          type: integer
          example: 422
        detail:
          type: string
        instance:
          type: string
          example: '/competencias/123'
    ProblemValidacion:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: precio_venta
                  message:
                    type: string
                    example: 'Debe ser mayor a 0'

    # --- Auth ---
...
    Decision:
      type: object
      properties:
        id:
          type: integer
          format: int64
        equipo_id:
          type: integer
          format: int64
        trimestre_id:
          type: integer
          format: int64
        registrado_por_usuario_id:
          type: integer
          format: int64
        prestamo_solicitado:
          type: integer
          format: int64
        dividendos_pagar:
          type: integer
          format: int64
        produccion_planificada:
          type: integer
          format: int64
        inversion_capacidad:
          type: integer
          format: int64
        precio_venta:
          type: integer
          format: int64
        inversion_marketing:
          type: integer
          format: int64
        contrataciones_netas:
          type: integer
        aumento_salarial_pct:
          type: number
          format: float
        inversion_capacitacion:
          type: integer
          format: int64
        inversion_id:
          type: integer
          format: int64
        estado:
          type: string
          enum: [BORRADOR, ENVIADA, PROCESADA]
        submitted_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []